create Propietario service, controller

+++ src/main/java/com/jcaido/demorender/services/codigoPostal/CodigoPostalServiceImpl.java
@@ -6,6 +6,7 @@ import com.jcaido.demorender.exceptions.BadRequestModificacionException;
 import com.jcaido.demorender.exceptions.ResourceNotFoundException;
 import com.jcaido.demorender.models.CodigoPostal;
 import com.jcaido.demorender.repositories.CodigoPostalRepository;
+import com.jcaido.demorender.services.propietario.PropietarioService;
 import org.modelmapper.ModelMapper;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -21,11 +22,13 @@ import java.util.Optional;
 public class CodigoPostalServiceImpl implements CodigoPostalService{
 
     private final CodigoPostalRepository codigoPostalRepository;
+    private final PropietarioService propietarioService;
     private final CodigoPostalValidacionesUniqueService codigoPostalValidacionesUniqueService;
     private  final ModelMapper modelMapper;
 
-    public CodigoPostalServiceImpl(CodigoPostalRepository codigoPostalRepository, CodigoPostalValidacionesUniqueService codigoPostalValidacionesUniqueService, ModelMapper modelMapper) {
+    public CodigoPostalServiceImpl(CodigoPostalRepository codigoPostalRepository, PropietarioService propietarioService, CodigoPostalValidacionesUniqueService codigoPostalValidacionesUniqueService, ModelMapper modelMapper) {
         this.codigoPostalRepository = codigoPostalRepository;
+        this.propietarioService = propietarioService;
         this.codigoPostalValidacionesUniqueService = codigoPostalValidacionesUniqueService;
         this.modelMapper = modelMapper;
     }
@@ -108,8 +111,8 @@ public class CodigoPostalServiceImpl implements CodigoPostalService{
         if (!codigoPostalRepository.existsById(id))
             throw new ResourceNotFoundException("Codigo Postal", "id", String.valueOf(id));
 
-        //if (propietarioService.obtenerPropietariosPorCodigoPostalHQL(id).size() > 0)
-        //    throw new ResponseStatusException(HttpStatus.CONFLICT, "Existen propietarios relacionados con ese codigo postal");
+        if (propietarioService.obtenerPropietariosPorCodigoPostalHQL(id).size() > 0)
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "Existen propietarios relacionados con ese codigo postal");
 
         codigoPostalRepository.deleteById(id);
